Move out the signature part into a new file
